[Release-1.5.0] Use FTRL in LR to improve performance

+++ spark-on-angel/core/src/test/java/com/tencent/angel/spark/pof/Zip2MapWithIndexFuncTest.java
@@ -33,7 +33,7 @@ public class Zip2MapWithIndexFuncTest implements Zip2MapWithIndexFunc {
   }
 
   @Override
-  public double call(int index, double value1, double value2) {
+  public double call(long index, double value1, double value2) {
     if (index == 0) {
       return value1 + value2 * value2;
     } else {
